chainlit_app/requirements.txt
# Chainlit App Dependencies
chainlit==1.0.101 
openai==1.3.7 
google-generativeai==0.3.1 
requests==2.31.0 
python-dotenv==1.0.0 
anthropic==0.5.0 # For Grok API (using Anthropic as placeholder) 
pydantic==2.4.2
